+++ tests/testthat/test-testGenesetNetwork.R
@@ -4,40 +4,42 @@ library(msigdb)
 test_that("geneset overlap computation works", {
   #empty collections
   nullgsc = GeneSetCollection(list())
-  expect_error(computeMsigOverlap(nullgsc, thresh = 0.15))
+  expect_error(computeMsigOverlap(nullgsc, thresh = 0.15, measure = 'jaccard'))
   
   #non-empty gscs but empty gene sets
   emptygsc = GeneSetCollection(GeneSet(setName = 'A', geneIdType = SymbolIdentifier()))
-  expect_error(computeMsigOverlap(emptygsc, thresh = 1))
+  expect_error(computeMsigOverlap(emptygsc, thresh = 1, measure = 'jaccard'))
   
   emptygsc = GSEABase::GeneSetCollection(GSEABase::GeneSet(setName = 'a'),
                                          GSEABase::GeneSet(c('1', '2'), setName = 'b'))
-  expect_error(computeMsigOverlap(emptygsc, thresh = 1))
-  expect_error(computeMsigOverlap(emptygsc[2], emptygsc[1], thresh = 1))
+  expect_error(computeMsigOverlap(emptygsc, thresh = 1, measure = 'jaccard'))
+  expect_error(computeMsigOverlap(emptygsc[2], emptygsc[1], thresh = 1, measure = 'jaccard'))
   
   #2 geneset
   data("hgsc")
   estgsc = hgsc[grep('ESTROGEN', hgsc)]
-  ov_jc = computeMsigOverlap(estgsc, thresh = 0.15)
+  ov_ari = computeMsigOverlap(estgsc, thresh = -Inf, measure = 'ari')
+  ov_jc = computeMsigOverlap(estgsc, thresh = 0.15, measure = 'jaccard')
   ov_oc = computeMsigOverlap(estgsc, thresh = 0.15, measure = 'ovlap')
   
+  expect_equal(round(ov_ari[, 3], 3), 0.091)
   expect_equal(round(ov_jc[, 3], 3), 0.338)
   expect_equal(round(ov_oc[, 3], 3), 0.505)
   
   #hallmark geneset
-  expect_error(nrow(computeMsigOverlap(hgsc, nullgsc, thresh = 0.15)))
-  expect_error(ncol(computeMsigOverlap(hgsc, nullgsc, thresh = 0.15)))
-  expect_equal(nrow(computeMsigOverlap(estgsc, thresh = 0.15)), 1)
-  expect_equal(nrow(computeMsigOverlap(hgsc, estgsc, thresh = 0.15)), 4)
-  expect_equal(nrow(computeMsigOverlap(hgsc, thresh = 0.15)), 6)
-  expect_equal(nrow(computeMsigOverlap(hgsc, thresh = 0)), 1225)
-  expect_equal(nrow(computeMsigOverlap(hgsc, thresh = 1)), 0)
+  expect_error(nrow(computeMsigOverlap(hgsc, nullgsc, thresh = 0.15, measure = 'jaccard')))
+  expect_error(ncol(computeMsigOverlap(hgsc, nullgsc, thresh = 0.15, measure = 'jaccard')))
+  expect_equal(nrow(computeMsigOverlap(estgsc, thresh = 0.15, measure = 'jaccard')), 1)
+  expect_equal(nrow(computeMsigOverlap(hgsc, estgsc, thresh = 0.15, measure = 'jaccard')), 4)
+  expect_equal(nrow(computeMsigOverlap(hgsc, thresh = 0.15, measure = 'jaccard')), 6)
+  expect_equal(nrow(computeMsigOverlap(hgsc, thresh = 0, measure = 'jaccard')), 1225)
+  expect_equal(nrow(computeMsigOverlap(hgsc, thresh = 1, measure = 'jaccard')), 0)
 })
 
 test_that("overlap network computation works", {
   #hallmark geneset
   data("hgsc")
-  ov = computeMsigOverlap(hgsc, thresh = 0.15)
+  ov = computeMsigOverlap(hgsc, thresh = 0.15, measure = 'jaccard')
   ig = computeMsigNetwork(ov, hgsc)
   
   expect_s3_class(ig, 'igraph')
@@ -50,7 +52,7 @@ test_that("overlap network computation works", {
 test_that("plot functions work", {
   #hallmark geneset
   data("hgsc")
-  ov = computeMsigOverlap(hgsc, thresh = 0.15)
+  ov = computeMsigOverlap(hgsc, thresh = 0.15, measure = 'jaccard')
   ig = computeMsigNetwork(ov, hgsc)
   grps = as.list(igraph::V(ig)$name)
   names(grps) = 1:length(grps)
@@ -70,7 +72,7 @@ test_that("plot functions work", {
 test_that("geneset neighbourhood works", {
   #hallmark geneset
   data("hgsc")
-  ov = computeMsigOverlap(hgsc, thresh = 0.15)
+  ov = computeMsigOverlap(hgsc, thresh = 0.15, measure = 'jaccard')
   ig = computeMsigNetwork(ov, hgsc)
   
   expect_equal(getMsigNeighbour('HALLMARK_ESTROGEN_RESPONSE_LATE', ig), 'HALLMARK_ESTROGEN_RESPONSE_EARLY')